Added MainGUI class, refactored game logic
This commit makes notable improvements in the GUI and game logic. A MainGUI class is added in order to facilitate the map size selection feature. The game logic in Controller.java is also refined, particularly in terms of handling paths. Additionally, 'generateBoard' function is now shifted to Generator.java for better code structure and clarity.